added DDL and procedure for updating corpora
motivation:
when updating a corpus, its ID should be stable, so that e.g. shared
linkgs still work

logic:
when updating a corpus A with corpus B, all relevant fields from B are
copied to A and the "history" (i.e. the copying from B -> A) is recorded
in main.corpus_history.
if this is the first update of A, its state before the update is
recorded in the "initial_state" column of main.corpus_history